style: fix cpplint header guard and use uint64_t for timestamp
Align header guard with cpplint path convention and replace unsigned
long long in get_process_utilization_samples to satisfy runtime/int.

+++ src/multiprocess/multiprocess_utilization_watcher.c
@@ -123,7 +123,7 @@ unsigned int cuda_to_nvml_map(unsigned int cudadev){
 
 static nvmlReturn_t get_process_utilization_samples(
     nvmlDevice_t device,
-    unsigned long long last_seen,
+    uint64_t last_seen,
     nvmlProcessUtilizationSample_t *out,
     unsigned int *out_count) {
   unsigned int processes_num = *out_count;
